fix(151): improve wrapper for image nodes
