+++ packages/modules.editor/src/hooks/index.ts
@@ -2,3 +2,4 @@ export { useTextFormat } from './useTextFormat';
 export { useYjsContext } from './useYjsContext';
 export { useEditorActive } from './useEditorActive';
 export { useBlockMenuActions } from './useBlockMenuActions';
+export { useProtectedImage } from './useProtectedImage';
